chore: content switch tab  분리
